# Improve engine events to expose internal workflow