--- a/packages/adblocker-electron/src/index.ts
+++ b/packages/adblocker-electron/src/index.ts
@@ -185,6 +185,11 @@ export class ElectronBlocker extends FiltersEngine {
       getExtendedRules: true,
       getRulesFromHostname: true,
       getRulesFromDOM: false, // Only done on updates (see `onGetCosmeticFiltersUpdated`)
+
+      callerContext: {
+        frameId: event.frameId,
+        processId: event.processId,
+      },
     });
 
     if (active === false) {
@@ -232,6 +237,13 @@ export class ElectronBlocker extends FiltersEngine {
 
       // This will be done every time we get information about DOM mutation
       getRulesFromDOM: true,
+
+      callerContext: {
+        frameId: event.frameId,
+        processId: event.processId,
+
+        lifecycle: msg.lifecycle,
+      },
     });
 
     if (active === false) {