--- a/packages/adblocker-webextension-example/background.ts
+++ b/packages/adblocker-webextension-example/background.ts
@@ -72,8 +72,8 @@ WebExtensionBlocker.fromLists(fetch, fullLists, {
     console.log('redirect', request.url, result);
   });
 
-  blocker.on('csp-injected', (request: Request) => {
-    console.log('csp', request.url);
+  blocker.on('csp-injected', (request: Request, csps: string) => {
+    console.log('csp', request.url, csps);
   });
 
   blocker.on('script-injected', (script: string, url: string) => {
@@ -84,9 +84,11 @@ WebExtensionBlocker.fromLists(fetch, fullLists, {
     console.log('style', url, style.length);
   });
 
-  blocker.on('html-filtered', (htmlSelectors: HTMLSelector[]) => {
-    console.log('html selectors', htmlSelectors);
+  blocker.on('html-filtered', (htmlSelectors: HTMLSelector[], url: string) => {
+    console.log('html selectors', htmlSelectors, url);
   });
+
+  blocker.on('filter-matched', console.log.bind(console, 'filter-matched'));
 
   console.log('Ready to roll!');
 });