--- a/packages/adblocker-webextension/src/index.ts
+++ b/packages/adblocker-webextension/src/index.ts
@@ -347,6 +347,11 @@ export class WebExtensionBlocker extends FiltersEngine {
       getExtendedRules: false,
       getRulesFromDOM: false,
       getRulesFromHostname: true,
+
+      callerContext: {
+        tabId: details.tabId,
+        frameId: details.frameId,
+      },
     });
     if (active === false) {
       return;
@@ -460,6 +465,11 @@ export class WebExtensionBlocker extends FiltersEngine {
         getExtendedRules: false,
         getRulesFromDOM: false,
         getRulesFromHostname: false,
+
+        callerContext: {
+          tabId: sender.tab?.id,
+          frameId: sender.frameId,
+        },
       });
 
       if (active === false) {
@@ -497,6 +507,11 @@ export class WebExtensionBlocker extends FiltersEngine {
 
         // This will be done every time we get information about DOM mutation
         getRulesFromDOM: msg.lifecycle === 'dom-update',
+
+        callerContext: {
+          tabId: sender.tab?.id,
+          frameId: sender.frameId,
+        },
       });
 
       if (active === false) {