fixes https://github.com/ghostery/adblocker/issues/3876