@seia-soto 2024-04-05
@remusao I'd like to get a feedback about the way passing "context" to matching functions. It might help a lot for extension developers often used to involve `tabId`.

@chrmod This is actually an off-topic but if we involve `devTools` API, we can avoid involving `tabId`. The devtools apis are already provided in scope of the page, so we don't need to take care about the tabs.

Passing an external context can solve many complex situations but I'd like to go simple as possible. For example, we could decorate the passed `Request` object and many other things. Also, we're already passing `request` as an argument for network filter matches.