@seia-soto 2024-04-30
In the [last changes](https://github.com/ghostery/adblocker/pull/3881/commits/0a9dc267481f42ab4938776f48f5595bc5553f71), the events were narrowed down to two: `filter-matched` and `filter-excluded`. Note that this idea was built top on the idea of adblocker debugger ui implementation.

Also, [the commit](https://github.com/ghostery/adblocker/pull/3881/commits/0a9dc267481f42ab4938776f48f5595bc5553f71) addresses the followings:
- Proper naming of `EngineEventContext` (renamed to `MatchingContext`): this variable name is expected to cause a lot of confusions.
- `MatchingContext` was updated to the union type with a proper enum typing to support the sole and integrated event kinds: `filter-matched` and `filter-excluded`.
- The additional argument can be passed to methods was renamed to `reference` instead of `context` to reduce confusion.