@seia-soto 2024-06-10
Updated benchmarking.

Skipping: serialization, list parsing, and initialization time - as these are not affected by this PR

For matched filters (94k samples), the maximum matching time increased a lot: from 1,770.5 μs to 4,453.25 μs. This can be seen as a big deal but if we look at the average: from 6.62 μs to 6.906 μs - , it shows that the processing time grows by following the linear graph.

For exception + unmatched filters (139k samples), the maximum matching time decreased: from 5,513.208 μs to 3,402.334 μs. This is because this PR dropped maximum 3 comparisons of if statements for exceptions if there's no listeners.

For the overall result (94k+139k samples), see the following:
- min: from 1.208 μs to 1.292 μs
- max: from 5,513.208 μs to 4,453.25 μs
- avg: from 7.846 μs to 8.183 μs

<details>
<summary>Full logs</summary>

```
(seia-soto:emit-cosmetic-filter-matches)

Avg serialization time (100 samples): 254.492 μs
Avg deserialization time (100 samples): 32.152 μs
Serialized size: 1,866 KiB
List parsing time: 147,284.666 μs
Initialization time: 21,659.791 μs

Total requests: 233,347
Total match: 94,294
Total no match: 139,053

Number of samples: 94,294
Min match: 1.292 μs
Max match: 4,453.25 μs
Avg match: 6.906 μs

Number of samples: 139,053
Min no match: 1.375 μs
Max no match: 3,402.334 μs
Avg no match: 9.048 μs

Number of samples: 233,347
Min (total): 1.292 μs
Max (total): 4,453.25 μs
Avg (total): 8.183 μs
```

```
(ghostery:master)

Avg serialization time (100 samples): 256.709 μs
Avg deserialization time (100 samples): 32.246 μs
Serialized size: 1,866 KiB
List parsing time: 151,396.167 μs
Initialization time: 23,166.625 μs

Total requests: 233,347
Total match: 94,294
Total no match: 139,053

Number of samples: 94,294
Min match: 1.208 μs
Max match: 1,770.5 μs
Avg match: 6.62 μs

Number of samples: 139,053
Min no match: 1.334 μs
Max no match: 5,513.208 μs
Avg no match: 8.678 μs

Number of samples: 233,347
Min (total): 1.208 μs
Max (total): 5,513.208 μs
Avg (total): 7.846 μs
```

</details>